requirements.txt
# 📋 requirements.txt - Dipendenze del progetto

# Telegram Bot
python-telegram-bot>=21.0

# Fuzzy Matching per squadre
rapidfuzz>=3.0.0

# HTTP Requests
requests>=2.31.0

# JSON per cache
python-dotenv>=1.0.0

# (Opzionali per Streamlit app)
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.0.0
